requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
WTForms==3.0.1
